fix: Refactor UEFI bootloader structure and implement new board support
