+++ bootloader/axloader/Cargo.toml
@@ -5,7 +5,7 @@ license.workspace = true
 name = "axloader"
 publish = false
 repository.workspace = true
-version.workspace = true
+version = "0.1.0"
 
 [[bin]]
 name = "axloader"
@@ -15,9 +15,7 @@ required-features = ["loader-bin"]
 [features]
 default = []
 loader-bin = []
-arch-x86_64 = []
-arch-loongarch64 = []
-board-asus-nuc15crh = ["arch-x86_64", "loader-bin"]
+board-asus-nuc15crh = ["loader-bin"]
 
 [dependencies]
 httpboot-protocol = { workspace = true }
