+++ bootloader/axloader/src/loader/boards/asus_nuc15crh.rs
